Added an in-project maven repo to include a third-party jar
The prowl-api jar is not available from maven central. The first
alternative was to install the jar manually in one's local maven reop
using the mvn install:install-file (as documented on
https://maven.apache.org/guides/mini/guide-3rd-party-jars-local.html).

However, this manual step can be avoided by creating an in-project maven
repo which has this missing dependency and using that in the pom
instead.