Fix verbose logging by excluding other sources for SLF4J bindings
The log output was verbose enough to affect the performance of the
scrapers. This was due to multiple SLF4J bindings on the classpath
resulting in log4j.properties not having an effect. This was fixed by
excluding other sources of SLF4J bindings to setup the log system
properly.

Change-Id: I39fac8ee26b146548ef4f8ec4953bd3807cb4ff7